[ARM Auto SignOff] Rename from preview to GA
