+++ .github/workflows/arm-incremental-typespec.yaml
@@ -1,4 +1,4 @@
-name: ARM Incremental TypeSpec (Preview)
+name: ARM Incremental TypeSpec
 
 on:
   pull_request:
@@ -13,8 +13,8 @@ permissions:
   contents: read
 
 jobs:
-  arm-incremental-typespec-preview:
-    name: ARM Incremental TypeSpec (Preview)
+  arm-incremental-typespec:
+    name: ARM Incremental TypeSpec
 
     runs-on: ubuntu-24.04
 
@@ -37,16 +37,16 @@ jobs:
           # "--omit dev": not needed at dev time, improves performance
           install-command: "npm ci --no-audit --omit dev"
           working-directory: ./.github
-    
+
       # Output is "true" if PR contains only incremental changes to an existing TypeSpec RP
       - id: incremental-typespec
-        name: ARM Incremental TypeSpec (Preview)
+        name: ARM Incremental TypeSpec
         uses: actions/github-script@v7
         with:
           result-encoding: string
           script: |
             const { default: incrementalTypeSpec } =
-              await import('${{ github.workspace }}/.github/workflows/src/arm-incremental-typespec-preview.js');
+              await import('${{ github.workspace }}/.github/workflows/src/arm-incremental-typespec.js');
             return await incrementalTypeSpec({ github, context, core });
 
       - name: Upload artifact with results
