+++ .github/workflows/update-labels.yaml
@@ -7,7 +7,12 @@ on:
   # If an upstream workflow if completed, get only the artifacts from that workflow, and update labels
   workflow_run:
     workflows:
-      ["ARM Auto SignOff (Preview)", "SDK Breaking Change Labels (Preview)", "SDK Suppressions", "TypeSpec Requirement"]
+      [
+        "ARM Auto SignOff",
+        "SDK Breaking Change Labels",
+        "SDK Suppressions",
+        "TypeSpec Requirement",
+      ]
     types: [completed]
   workflow_dispatch:
     inputs:
